Upload All the Files
New repository because there were some private keys and private codes. Small amounts of the code were developed by ironminer45 and 1AndOnlyPika both can be found on github. I didn't develop anything in the nodemodules folder I just thought it would be helpful for anyone using this to know what was required, along with what versions it was developed with. Credit for those nodemodules should be inside them in the licence file.